requirements.txt
pandas==1.5.3
numpy==1.23.5
tqdm==4.65.0
faiss-cpu==1.7.3
nltk==3.8.1
googletrans==3.1.0a0
sentence-transformers==2.2.2
transformers==4.25.1
memory-profiler==0.61.0
swifter==1.3.4
pyarrow==10.0.1
regex==2022.10.31
fuzzywuzzy==0.18.0
matplotlib==3.7.1
python-Levenshtein==0.20.9
ipywidgets==7.5.0